cast graph size to int to make visual studio stop complaining at me

+++ Aquila/aquila.cpp
@@ -49,7 +49,7 @@ namespace aquila {
 	}
 
 	Graph::Graph() {
-		id = graphs.size();
+		id = (int)graphs.size();
 	}
 
 	Graph::~Graph() {
